move functions to resolver.go
